+++ packages/components/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @equinor/mad-components
 
+## 0.11.2
+
+### Patch Changes
+
+-   7eae144: Added useBreakpoint hook to use for apps and components that have specific requirements
+    or styles based on device size
+
+## 0.11.1
+
+### Patch Changes
+
+-   d986ab2: Add `additionalSurface` prop to `Cell`component that makes it possible for cell to have
+    two unique `onPress` events.
+
 ## 0.11.0
 
 ### Minor Changes