+++ packages/skia-draw/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @equinor/react-native-skia-draw
 
+## 0.3.9
+
+### Patch Changes
+
+-   Updated dependencies [7eae144]
+    -   @equinor/mad-components@0.11.2
+
+## 0.3.8
+
+### Patch Changes
+
+-   Updated dependencies [d986ab2]
+    -   @equinor/mad-components@0.11.1
+
 ## 0.3.7
 
 ### Patch Changes